[bd-a924k] AR + Payments: amount_cents i32 to i64 (SageDesert)
AR 5.0.0→6.0.0, Payments 4.0.0→5.0.0. Migration widens columns to BIGINT.

+++ modules/ar/src/consumer_tasks.rs
@@ -199,7 +199,7 @@ pub async fn handle_payment_succeeded(
     if let Some(row) = row {
         let app_id: String = row.get("app_id");
         let customer_id: i32 = row.get("ar_customer_id");
-        let amount_cents: i32 = row.get("amount_cents");
+        let amount_cents: i64 = row.get("amount_cents");
         let currency: String = row.get("currency");
         let created_at = row.get("created_at");
         let due_at = row.get("due_at");
