+++ modules/ar/src/models/invoice.rs
@@ -27,7 +27,7 @@ pub struct Invoice {
     pub ar_customer_id: i32,
     pub subscription_id: Option<i32>,
     pub status: String,
-    pub amount_cents: i32,
+    pub amount_cents: i64,
     pub currency: String,
     pub due_at: Option<NaiveDateTime>,
     pub paid_at: Option<NaiveDateTime>,
@@ -51,7 +51,7 @@ pub struct CreateInvoiceRequest {
     pub ar_customer_id: i32,
     pub subscription_id: Option<i32>,
     pub status: Option<String>,
-    pub amount_cents: i32,
+    pub amount_cents: i64,
     pub currency: Option<String>,
     pub due_at: Option<NaiveDateTime>,
     pub metadata: Option<JsonValue>,
@@ -68,7 +68,7 @@ pub struct CreateInvoiceRequest {
 #[derive(Debug, Deserialize, ToSchema)]
 pub struct UpdateInvoiceRequest {
     pub status: Option<String>,
-    pub amount_cents: Option<i32>,
+    pub amount_cents: Option<i64>,
     pub due_at: Option<NaiveDateTime>,
     pub metadata: Option<JsonValue>,
 }
