+++ modules/payments/src/models.rs
@@ -11,7 +11,7 @@ use uuid::Uuid;
 pub struct PaymentCollectionRequestedPayload {
     pub invoice_id: String,
     pub customer_id: String,
-    pub amount_minor: i32,
+    pub amount_minor: i64,
     pub currency: String,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub payment_method_id: Option<String>,
@@ -27,7 +27,7 @@ pub struct PaymentSucceededPayload {
     pub payment_id: String,
     pub invoice_id: String,
     pub ar_customer_id: String,
-    pub amount_minor: i32,
+    pub amount_minor: i64,
     pub currency: String,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub processor_payment_id: Option<String>,
@@ -41,7 +41,7 @@ pub struct PaymentFailedPayload {
     pub payment_id: String,
     pub invoice_id: String,
     pub ar_customer_id: String,
-    pub amount_minor: i32,
+    pub amount_minor: i64,
     pub currency: String,
     pub failure_code: String,
     #[serde(skip_serializing_if = "Option::is_none")]
